+++ packages/v2/gui/src/fakta/sykdom-og-opplæring/3-nødvendig-opplæring/NødvendigOpplæringFerdigvisning.tsx
@@ -1,16 +1,33 @@
-import type { OpplæringVurderingDto } from '@k9-sak-web/backend/k9sak/generated';
-import { BodyShort } from '@navikt/ds-react';
-import type { Period } from '@navikt/ft-utils';
+import {
+  OpplæringVurderingDtoAvslagsårsak,
+  OpplæringVurderingDtoResultat,
+  type OpplæringVurderingDto,
+} from '@k9-sak-web/backend/k9sak/generated';
+import { Alert, BodyShort } from '@navikt/ds-react';
 import { LabelledContent } from '../../../shared/labelled-content/LabelledContent';
 import { Lovreferanse } from '../../../shared/lovreferanse/Lovreferanse';
 import { VurdertAv } from '../../../shared/vurdert-av/VurdertAv';
-const NødvendigOpplæringFerdigvisning = ({ vurdering }: { vurdering: OpplæringVurderingDto & { perioder: Period[] } }) => {
+import { K9KodeverkoppslagContext } from '../../../kodeverk/oppslag/K9KodeverkoppslagContext';
+import { useContext } from 'react';
+const NødvendigOpplæringFerdigvisning = ({ vurdering }: { vurdering: OpplæringVurderingDto }) => {
+  if (vurdering.resultat === OpplæringVurderingDtoResultat.VURDERES_SOM_REISETID) {
+    return (
+      <Alert variant="info" size="small">
+        Dag eller periode er flyttet til vurdering av reisetid.
+      </Alert>
+    );
+  }
+
   return (
     <div className="flex flex-col gap-6">
       <LabelledContent
         label="Er nødvendig opplæring dokumentert med legeerklæring?"
         size="small"
-        content={<BodyShort size="small">{vurdering.dokumentertOpplæring ? 'Ja' : 'Nei'}</BodyShort>}
+        content={
+          <BodyShort size="small">
+            {vurdering.resultat === OpplæringVurderingDtoResultat.GODKJENT ? 'Ja' : 'Nei'}
+          </BodyShort>
+        }
       />
       <div>
         <LabelledContent
@@ -35,10 +52,29 @@ const NødvendigOpplæringFerdigvisning = ({ vurdering }: { vurdering: Opplærin
       <LabelledContent
         label="Har søker hatt opplæring som er nødvendig for å kunne ta seg av og behandle barnet?"
         size="small"
-        content={<BodyShort size="small">{vurdering.nødvendigOpplæring ? 'Ja' : 'Nei'}</BodyShort>}
+        content={
+          <BodyShort size="small">
+            {vurdering.resultat === OpplæringVurderingDtoResultat.GODKJENT ? 'Ja' : 'Nei'}
+          </BodyShort>
+        }
       />
+      {vurdering.avslagsårsak && vurdering.avslagsårsak !== OpplæringVurderingDtoAvslagsårsak.UDEFINERT && (
+        <LabelledContent
+          label="Avslagsårsak"
+          size="small"
+          content={<AvslagårsakTekst avslagsårsak={vurdering.avslagsårsak} />}
+        />
+      )}
     </div>
   );
+};
+
+const AvslagårsakTekst = ({ avslagsårsak }: { avslagsårsak: OpplæringVurderingDtoAvslagsårsak }) => {
+  const k9Kodeverkoppslag = useContext(K9KodeverkoppslagContext);
+  if (!avslagsårsak) {
+    return null;
+  }
+  return <BodyShort size="small">{(k9Kodeverkoppslag as any).k9sak.avslagsårsaker(avslagsårsak).navn}</BodyShort>;
 };
 
 export default NødvendigOpplæringFerdigvisning;
